refactor(agent): test production loop detector
