+++ TODOs.md
@@ -194,15 +194,15 @@ parity regressions likely.
 
 ## 9. Test the real shared logic instead of copied shims
 
-`test/run.js` duplicates pieces of `Agent` logic in `LoopDetectorShim` because
-`agent.js` imports browser-only modules. That means tests can pass while the real
-agent implementation drifts.
+**Status:** Partially resolved. Loop detection now lives in the browser-free
+`agent/loop-detector.js` module in both builds. `Agent` inherits that production
+class, `test/run.js` imports it directly, and a parity assertion keeps the
+Chrome and Firefox copies byte-identical.
 
 **Concrete next steps:**
-1. Move loop detection, coordinate-click bucketing, image budget sizing, and
-   other pure logic into browser-free modules.
-2. Import those modules directly from both `agent.js` and `test/run.js`.
-3. Add regression tests for the text tool-call parser and context trimming,
+1. Move image budget sizing and other remaining pure logic into browser-free
+   modules and import those modules directly from `agent.js` and `test/run.js`.
+2. Add regression tests for the text tool-call parser and context trimming,
    since both are high-impact agent reliability code.
 
 ---
