+++ src/chrome/ARCHITECTURE.md
@@ -41,6 +41,7 @@ src/chrome/
 │   ├── run-ui-journal.js       # Detached-run replay + streamed-text snapshots
 │   ├── agent/
 │   │   ├── agent.js            # Core agent loop + tool dispatch
+│   │   ├── loop-detector.js     # Browser-free loop detection, directly unit-tested
 │   │   ├── tools.js            # Tool schemas + system prompts
 │   │   ├── skills.js           # Settings skills + dynamic skill tool manifests
 │   │   ├── planner.js          # Plan-before-Act structured planner
@@ -562,6 +563,9 @@ OpenAI format → Anthropic blocks: system → separate `system` field; `assista
 ---
 
 ## Loop Detection
+
+The browser-free implementation lives in `agent/loop-detector.js`; `Agent`
+inherits it, and the unit suite imports the same production class directly.
 
 Three independent detectors, strongest action wins:
